Unify MessagingExceptionListener between clients and don't System.exit() there
Main reason is that I seem to be unable to intercept this System.exit from
selftests. Another reason is that it simplifies the code (less places
to exit) and that it does not seem to be necessary to exit there.
