+++ jmslib/src/main/java/com/redhat/mqe/lib/MessagingExceptionListener.java
@@ -1,0 +1,39 @@
+/*
+ * Copyright (c) 2018 Red Hat, Inc.
+ *
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements. See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.redhat.mqe.lib;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import javax.jms.ExceptionListener;
+import javax.jms.JMSException;
+
+/**
+ * MessagingExceptionListener is created for each connection made.
+ */
+public class MessagingExceptionListener implements ExceptionListener {
+    private Logger LOG = LoggerFactory.getLogger(ConnectionManager.class.getName());
+
+    @Override
+    public void onException(JMSException e) {
+        LOG.error("ExceptionListener error detected! \n{}\n{}", e.getMessage(), e.getCause());
+        e.printStackTrace();
+    }
+}
